consolidate packages
